+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserStorage {
     User addUser(User user);
 
-    User getUser(int id);
+    User getUser(long id);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     List<User> getUsers();
